+++ src/super/i-block/CHANGELOG.md
@@ -13,8 +13,8 @@ Changelog
 
 #### :rocket: New Feature
 
-* Added possibility to change method that will be used for transitions when router
-  synchronizes its state with the component's state by using syncRouterState
+* Added possibility to change the method that will be used for transitions when the router
+  synchronizes its state with the component's state by using `syncRouterState`
 
 ## v3.38.0 (2023-02-20)