+++ src/super/i-block/CHANGELOG.md
@@ -9,7 +9,7 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
-## v3.??.?? (2023-??-??)
+## v3.54.0 (2023-07-28)
 
 #### :rocket: New Feature